+++ services/notifications/src/simcore_service_notifications/repositories/_file_templates_repository.py
